requirements.txt
pad4pi
time
playsound
datetime
RPi.GPIO
